Refactor role and user management, add Role REST controller
Changed Role entity ID from Integer to Long and updated related repositories and methods. Replaced CrudRepository with JpaRepository for RoleRepository. Enhanced User entity with builder and all-args constructor, and updated UserRepository to use Long for role IDs. Updated AdminController to support user creation via DTO, including roles, interests, and related entities. Added a new RolRestController for CRUD operations on roles. Updated role descriptions in RoleSeeder to Spanish. Restricted user listing and update endpoints in UserRestController to SUPER_ADMIN role. Extended RegisterUserRequestDTO to support role IDs.

+++ src/main/java/com/project/demo/logic/entity/user/UserRepository.java
@@ -26,5 +26,5 @@ public interface UserRepository extends JpaRepository<User, Long>  {
         SELECT r.id FROM u.roles r
     )
 """)
-    Page<User> findAllExcludingUsersWithRoleId(@Param("roleId") Integer roleId, Pageable pageable);
+    Page<User> findAllExcludingUsersWithRoleId(@Param("roleId") Long roleId, Pageable pageable);
 }
